requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
Babel>=1.3

oslo.config>=3.7.0  # Apache-2.0
oslo.i18n>=2.1.0  # Apache-2.0
oslo.log>=1.14.0  # Apache-2.0
oslo.serialization>=1.10.0              # Apache-2.0
oslo.utils>=3.5.0                       # Apache-2.0

python-neutronclient

oslo.vmware>=0.5
